requirements.txt
colorama==0.4.5
playsound==1.3.0
pypng==0.20220715.0
PyQRCode==1.2.1
